remove unnecessary configs
