+++ packages/web/src/lib/world/configs.ts
@@ -1,5 +1,3 @@
-import type { Sprite } from "pixi.js";
-
 export const SPRITE_CONFIG = {
   ANT: { scale: 1.8, anchor: { x: 0.5, y: 0 } },
   FOOD: { scale: 1.5, anchor: { x: 0.5, y: -1 } },
@@ -12,38 +10,8 @@ export const ANIMATION_CONFIG = {
   hideSpriteRadius: 60,
 } as const;
 
-export const BACKGROUND_CONFIG = {
-  tileSize: 16,
-  terrainTileNames: [
-    "grass-dark-0",
-    "grass-dark-1",
-    "grass-dark-2",
-    "grass-dark-3",
-  ],
-  grassTiles: ["grass-plain", "grass-2", "grass-1", "grass-3"],
-  grassWeights: [0.75, 0.85, 0.95, 1.0],
-} as const;
-
-export const BOULDER_CONFIG = {
-  count: 6,
-  textures: ["/background/boulder-1.png", "/background/boulder-2.png"],
-  minScale: 0.4,
-  maxScale: 0.8,
-} as const;
-
-export const FOOD_SOURCE_CONFIG = {
-  maxAmount: 100,
-} as const;
-
-export const LAYERS = {
+export const LAYER_INDEX = {
   BACKGROUND: 0,
   STATIC_OBJECTS: 1,
   ENTITIES: 2,
 } as const;
-
-export type AntSprite = {
-  sprite: Sprite;
-  previousPosition: { x: number; y: number };
-  direction: string;
-  animationFrame: number;
-};
