+++ packages/web/src/lib/world/sprite.ts
@@ -1,6 +1,13 @@
 import { Sprite, type Texture } from "pixi.js";
 import type { SpriteConfig } from "./configs";
 
+export type AntSprite = {
+  sprite: Sprite;
+  previousPosition: { x: number; y: number };
+  direction: string;
+  animationFrame: number;
+};
+
 export const createSpriteWithConfig = (
   texture: Texture,
   config: SpriteConfig,
